split the existing tools into separate modules
