+++ tools/internal/contract-state-importer/importer-cli/src/commands/initialise_mixnet_vesting_with_states.rs
@@ -1,0 +1,4 @@
+// Copyright 2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: GPL-3.0-only
+
+use clap::ArgGroup;
